refactor: useCardManage 분리
