+++ src/contexts/PaymentContextProvider.js
@@ -1,38 +1,11 @@
 import React, { createContext, useState } from 'react';
-import { getCardId, getCardNickName } from '../utils/getCardInfo';
-
-const registeredCards = {};
+import { useCardManage } from '../hooks/useCardManage';
 
 export const PaymentContext = createContext();
 
 export const PaymentContextProvider = ({ children }) => {
   const [currentPage, setCurrentPage] = useState('cardList');
-  const [cards, setCards] = useState(registeredCards);
-  const [cardId, setCardId] = useState(null);
-
-  const registerCard = (card) => {
-    const cardId = getCardId();
-    const nickName = getCardNickName(card.company, card.owner);
-
-    setCards((cards) => ({ ...cards, [cardId]: { ...card, nickName } }));
-    setCardId(cardId);
-  };
-
-  const updateCard = (nickName) => {
-    const card = cards[cardId];
-
-    setCards((cards) => ({ ...cards, [cardId]: { ...card, nickName } }));
-    setCardId(null);
-  };
-
-  const deleteCard = () => {
-    setCards((cards) => {
-      delete cards[cardId];
-
-      return { ...cards };
-    });
-    setCardId(null);
-  };
+  const { cards, cardId, setCardId, registerCard, updateCard, deleteCard } = useCardManage();
 
   return (
     <PaymentContext.Provider
